+++ docs/source/changelog.rst
@@ -13,6 +13,8 @@ ChangeLog
 * The ``CA_DIGEST_ALGORITHM`` was removed. Use :ref:`settings-ca-default-signature-hash-algorithm` instead.
 * The ``CA_DEFAULT_ECC_CURVE`` was removed. Use :ref:`settings-ca-default-elliptic-curve` instead.
 * Support for non-standard algorithm names for the ``--algorithm`` argument was removed.
+* Support for non-standard elliptic curve names via the ``--elliptic-curve`` argument was removed.
+* The old alias ``--ecc-curve`` for ``--elliptic-curve`` was removed.
 
 .. _changelog-1.24.0: